Uploaded Course Material for 'Initiation à la Recherche Scientifique
This commit includes the upload of the course materials for "Initiation à la Recherche Scientifique." The content covers introductory elements, the strategic outlook of Morocco, scientific literature and its components, editorial processes, scientific activities, and research regulations. Contributors are invited to review, provide feedback, and contribute to the enhancement of the course content. This upload marks a significant step in making the course resources available for collaboration and learning.

Co-Authored-By: tingeek <153770178+tingeek@users.noreply.github.com>
Co-Authored-By: BOUHLALI Abdelfattah <83462650+bydevmar@users.noreply.github.com>